Create 1850B.cpp
Codeforces Round 886 (Div. 4)
B. Ten Words of Wisdom

In the game show "Ten Words of Wisdom", there are n participants numbered from 1 to 𝑛, each of whom submits one response. The 𝑖-th response is 𝑎𝑖 words long and has quality 𝑏𝑖. No two responses have the same quality, and at least one response has length at most 10.The winner of the show is the response which has the highest quality out of all responses that are not longer than 10 words. Which response is the winner?