Create 1777A.cpp
Codeforces Round 845 (Div. 2) and ByteRace 2023
A. Everybody Likes Good Arrays!

An array 𝑎 is good if for all pairs of adjacent elements, 𝑎𝑖 and 𝑎𝑖+1
(1≤𝑖<𝑛) are of different parity. Note that an array of size 1 is trivially good.
You are given an array of size n.In one operation you can select any pair of adjacent elements in which both elements are of the same parity, delete them, and insert their product in the same position.Find the minimum number of operations to form a good array.