Create 1734B.c
B. Bright, Nice, Brilliant

There is a pyramid which consists of 𝑛 floors. The floors are numbered from top to bottom in increasing order. In the pyramid, the 𝑖-th floor consists of 𝑖 rooms.

Denote the 𝑗-th room on the 𝑖-th floor as (𝑖,𝑗). For all positive integers 𝑖 and 𝑗 such that 1≤𝑗≤𝑖<𝑛, there are 2 one-way staircases which lead from (𝑖,𝑗) to (𝑖+1,𝑗) and from (𝑖,𝑗) to (𝑖+1,𝑗+1) respectively.

In each room you can either put a torch or leave it empty. Define the brightness of a room (𝑖,𝑗) to be the number of rooms with a torch from which you can reach the room (𝑖,𝑗) through a non-negative number of staircases.
The pyramid is called nice if and only if for all floors, all rooms in the floor have the same brightness.

Define the brilliance of a nice pyramid to be the sum of brightness over the rooms (1,1), (2,1), (3,1), ..., (𝑛,1).Find an arrangement of torches in the pyramid, such that the resulting pyramid is nice and its brilliance is maximized.
We can show that an answer always exists. If there are multiple answers, output any one of them.