Create 1730A.cpp
A. Planets
One day, Vogons wanted to build a new hyperspace highway through a distant system with 𝑛 planets. The 𝑖-th planet is on the orbit 𝑎𝑖, there could be multiple planets on the same orbit. It's a pity that all the planets are on the way and need to be destructed.Vogons have two machines to do that.
The first machine in one operation can destroy any planet at cost of 1 Triganic Pu.The second machine in one operation can destroy all planets on a single orbit in this system at the cost of 𝑐 Triganic Pus.Vogons can use each machine as many times as they want.
Vogons are very greedy, so they want to destroy all planets with minimum amount of money spent. Can you help them to know the minimum cost of this project?