Create 1686A.cpp
A. Everything Everywhere All But One

You are given an array of 𝑛 integers 𝑎1,𝑎2,…,𝑎𝑛. After you watched the amazing film "Everything Everywhere All At Once", you came up with the following operation.In one operation, you choose 𝑛−1 elements of the array and replace each of them with their arithmetic mean (which doesn't have to be an integer). For example, from the array [1,2,3,1] we can get the array [2,2,2,1], if we choose the first three elements, or we can get the array [43,43,3,43], if we choose all elements except the third.
Is it possible to make all elements of the array equal by performing a finite number of such operations?