Create 1848A.cpp
A. Vika and Her Friends

Vika and her friends went shopping in a mall, which can be represented as a rectangular grid of rooms with sides of length n and m. Each room has coordinates (𝑎,𝑏), where 1≤𝑎≤𝑛,1≤𝑏≤𝑚. Thus we call a hall with coordinates (𝑐,𝑑) a neighbouring for it if  |𝑎−𝑐|+|𝑏−𝑑|=1.Tired of empty fashion talks, Vika decided to sneak away unnoticed. But since she hasn't had a chance to visit one of the shops yet, she doesn't want to leave the mall. After a while, her friends noticed Vika's disappearance and started looking for her.

Currently, Vika is in a room with coordinates (𝑥,𝑦), and her 𝑘 friends are in rooms with coordinates (𝑥1,𝑦1), (𝑥2,𝑦2), ... ,(𝑥𝑘,𝑦𝑘), respectively. The coordinates can coincide. Note that all the girls must move to the neighbouring rooms.
Every minute, first Vika moves to one of the adjacent to the side rooms of her choice, and then each friend (seeing Vika's choice) also chooses one of the adjacent rooms to move to.

If at the end of the minute (that is, after all the girls have moved on to the neighbouring rooms) at least one friend is in the same room as Vika, she is caught and all the other friends are called.

Tell us, can Vika run away from her annoying friends forever, or will she have to continue listening to empty fashion talks after some time?