Create 1823B.cpp
B. Sort with Step

Let's define a permutation of length 𝑛 as an array 𝑝 of length 𝑛, which contains every number from 1 to 𝑛 exactly once.
You are given a permutation 𝑝1,𝑝2,…,𝑝𝑛 and a number k. You need to sort this permutation in the ascending order. In order to do it, you can repeat the following operation any number of times (possibly, zero):
pick two elements of the permutation 𝑝𝑖 and 𝑝𝑗 such that |𝑖−𝑗|=𝑘, and swap them.
Unfortunately, some permutations can't be sorted with some fixed numbers 𝑘. For example, it's impossible to sort [2,4,3,1] with 𝑘=2.That's why, before starting the sorting, you can make at most one preliminary exchange:
choose any pair 𝑝𝑖 and 𝑝𝑗 and swap them.
Your task is to:
check whether is it possible to sort the permutation without any preliminary exchanges,
if it's not, check, whether is it possible to sort the permutation using exactly one preliminary exchange.
For example, if 𝑘=2 and permutation is [2,4,3,1], then you can make a preliminary exchange of 𝑝1 and 𝑝4, which will produce permutation [1,4,3,2], which is possible to sort with given 𝑘.