Create 1806B.cpp
B. Mex Master

You are given an array 𝑎 of length 𝑛. The score of 𝑎 is the MEX†
of [𝑎1+𝑎2,𝑎2+𝑎3,…,𝑎𝑛−1+𝑎𝑛]. Find the minimum score of 𝑎 if you are allowed to rearrange elements of 𝑎 in any order. Note that you are not required to construct the array 𝑎 that achieves the minimum score.
The MEX (minimum excluded) of an array is the smallest non-negative integer that does not belong to the array. For instance:

The MEX of [2,2,1] is 0, because 0 does not belong to the array.
The MEX of [3,1,0,1] is 2, because 0 and 1 belong to the array, but 2 does not.
The MEX of [0,3,1,2] is 4, because 0,1, 2, and 3 belong to the array, but 4 does not.