Create 1768B.cppp
B. Quick Sort

You are given a permutation 𝑝 of length 𝑛 and a positive integer 𝑘≤𝑛.In one operation, you:
Choose 𝑘 distinct elements 𝑝𝑖1,𝑝𝑖2,…,𝑝𝑖𝑘.Remove them and then add them sorted in increasing order to the end of the permutation.
For example, if 𝑝=[2,5,1,3,4] and 𝑘=2 and you choose 5 and 3 as the elements for the operation, then [2,5,1,3,4]→[2,1,4,3,5]
Find the minimum number of operations needed to sort the permutation in increasing order. It can be proven that it is always possible to do so.
A permutation of length 𝑛 is an array consisting of 𝑛 distinct integers from 1 to 𝑛 in arbitrary order. For example, [2,3,1,5,4] is a permutation, but [1,2,2] is not a permutation (2 appears twice in the array), and [1,3,4] is also not a permutation (𝑛=3 but there is 4 in the array).