Create 1675B.cpp
B. Make It Increasing

Given 𝑛 integers 𝑎1,𝑎2,…,𝑎𝑛. You can perform the following operation on them:
select any element 𝑎𝑖(1≤𝑖≤𝑛) and divide it by 2(round down). In other words, you can replace any selected element 𝑎𝑖 with the value ⌊𝑎𝑖/2⌋ (where ⌊𝑥⌋ is – round down the real number 𝑥). Output the minimum number of operations that must be done for a sequence of integers to become strictly increasing (that is, for the condition 𝑎1<𝑎2<⋯<𝑎𝑛 to be satisfied). Or determine that it is impossible to obtain such a sequence. Note that elements cannot be swapped. The only possible operation is described above.
For example, let 𝑛=3 and a sequence of numbers [3,6,5] be given. Then it is enough to perform two operations on it:
Write the number ⌊6/2⌋=3 instead of the number 𝑎2=6 and get the sequence [3,3,5].Then replace 𝑎1=3 with ⌊3/2⌋=1 and get the sequence [1,3,5].The resulting sequence is strictly increasing because 1<3<5.